+++ sensor-simulator/src/modules/simulator/simulatorService.js
@@ -1,41 +1,39 @@
-// src/modules/simulator/simulatorService.js
 const simulatorRepository = require('./simulatorRepository');
+const simulator = require('../../devices/simulator');
 
 function startDevice(deviceId) {
-  if (simulatorRepository.isDeviceRunning(deviceId)) {
-    return `${deviceId} 는 이미 존재합니다.`;
-  }
-  simulatorRepository.startDevice(deviceId);
+  simulator.startDevice(deviceId);
   return `${deviceId}의 시뮬레이션을 시작합니다.`;
 }
 
 async function restartDevice(deviceId) {
-  if (!simulatorRepository.isDeviceRunning(deviceId)) {
-    throw new Error(`${deviceId}가 존재하지 않습니다.`);
-  }
-  simulatorRepository.stopDevice(deviceId);
+  simulator.stopDevice(deviceId);
   await simulatorRepository.deleteStatusAndAlertsByDeviceId(deviceId);
-  simulatorRepository.startDevice(deviceId);
+  simulator.startDevice(deviceId);
   return `${deviceId}의 시뮬레이션을 재시작합니다.`;
 }
 
 function getRunningDevices() {
-  return simulatorRepository.getAllRunningDeviceIds();
+  return simulator.getAllRunningDeviceIds();
 }
 
 async function deleteDevice(deviceId) {
-  if (!simulatorRepository.isDeviceRunning(deviceId)) {
-    throw new Error(`${deviceId}가 존재하지 않습니다.`);
-  }
-  simulatorRepository.stopDevice(deviceId);
+  simulator.stopDevice(deviceId);
   await simulatorRepository.deleteStatusAndAlertsByDeviceId(deviceId);
   return `${deviceId} 시뮬레이션이 삭제되었습니다.`;
 }
 
 async function deleteAllDevices() {
-  simulatorRepository.stopAllDevices();
+  const runningDevices = simulator.getAllRunningDeviceIds();
+
+  if (runningDevices.length === 0) {
+    return '현재 실행 중인 장비가 없습니다.';
+  }
+
+  simulator.stopAllDevices();
   await simulatorRepository.deleteAllStatusAndAlerts();
-  return '모든 시뮬레이션이 삭제되었습니다.';
+
+  return `총 ${runningDevices.length}개의 장비가 중지 및 삭제되었습니다.`;
 }
 
 module.exports = {
